+++ src/component/Layout.tsx
@@ -1,10 +1,10 @@
 import Head from "next/head";
 import { ReactNode, FC } from "react";
 
-import Header from "src/component/Header";
-import Footer from "src/component/Footer";
+import { Header } from "src/component/Header";
+import { Footer } from "src/component/Footer";
 
-const Layout: FC<{ children: ReactNode; title: string }> = (props) => {
+export const Layout: FC<{ children: ReactNode; title: string }> = (props) => {
   return (
     <>
       <Head>
@@ -23,5 +23,3 @@ const Layout: FC<{ children: ReactNode; title: string }> = (props) => {
     </>
   );
 };
-
-export default Layout;
