+++ src/pages-component/Blog/index.ts
@@ -1,3 +1,1 @@
-import Blog from "./Blog";
-
-export default Blog;
+export { Blog } from "./page";
