+++ src/pages/about-page.tsx
@@ -1,8 +1,8 @@
 import { Container, Text } from "@mantine/core";
 import { FC } from "react";
 
-import ContentLayout from "src/component/ContentLayout";
-import Layout from "src/component/Layout";
+import { ContentLayout } from "src/component/ContentLayout";
+import { Layout } from "src/component/Layout";
 
 const AboutPage: FC = () => {
   return (
