+++ src/pages/portfolio-page.tsx
@@ -1,8 +1,8 @@
 import { Container } from "@mantine/core";
 import { NextPage } from "next";
 
-import Layout from "src/component/Layout";
-import Portfolio from "src/pages-component/Portfolio";
+import { Layout } from "src/component/Layout";
+import { Portfolio } from "src/pages-component/portfolio";
 
 const PortfolioPage: NextPage = () => {
   return (
