+++ src/types/Portfolio.ts
@@ -1,4 +1,4 @@
-export type Portfolio = {
+export type PortfolioData = {
   id: number;
   image: string;
   title: string;
